+++ src/components/help.rs
@@ -22,7 +22,7 @@ const GENERAL_DOCS: &[&[&str; 2]; 9] = &[
     &["Focus settings", "Tab"],
     &["Change setting", "Enter"],
 ];
-const SCOREBOARD_DOCS: &[&[&str; 2]; 9] = &[
+const SCOREBOARD_DOCS: &[&[&str; 2]; 10] = &[
     &["Scoreboard", "1"],
     &["Move down", "j/↓"],
     &["Move up", "k/↑"],
@@ -32,8 +32,9 @@ const SCOREBOARD_DOCS: &[&[&str; 2]; 9] = &[
     &["Scroll boxscore down", "Shift + j/↓"],
     &["Scroll boxscore up", "Shift + k/↑"],
     &["Toggle win probability", "w"],
+    &["Toggle sort mode", "s"],
 ];
-const GAMEDAY_DOCS: &[&[&str; 2]; 12] = &[
+const GAMEDAY_DOCS: &[&[&str; 2]; 13] = &[
     &["Gameday", "2"],
     &["Toggle game info", "i"],
     &["Toggle pitches", "p"],
@@ -46,6 +47,7 @@ const GAMEDAY_DOCS: &[&[&str; 2]; 12] = &[
     &["Move up at bat", "k/↑"],
     &["Go to live at bat", "l"],
     &["Go to first at bat", "s"],
+    &["Toggle scoring plays only", "!"],
 ];
 const STATS_DOCS: &[&[&str; 2]; 17] = &[
     &["Stats", "3"],